clean up settings and add accelerom data

+++ ControlForm.cs
@@ -85,43 +85,54 @@ namespace HarmoniaRemote
                             int intMetadataID = int.Parse(arrayValue[0].ToString());
                             String strValue = arrayValue[1].ToString();
 
-                            if (intMetadataID == 13) { SetControlText(this.meta_id_13,strValue); }//rtc
-                            if (intMetadataID == 4) {
-                                //this is the state - check for alarm
-                                SetControlText(this.meta_id_4,strValue);
-                                
+                            //search for the matching control and write the value to it
+                            string strControlName = "meta_id_" + intMetadataID.ToString();
+                            if (this.Controls.ContainsKey(strControlName))
+                            {
+                                Control ctr = this.Controls[strControlName];
+                                SetControlText(ctr, strValue);
                             }
-                            if (intMetadataID == 1) { SetControlText(this.meta_id_1,strValue); }
+
+                            //if (intMetadataID == 13) { SetControlText(this.meta_id_13,strValue); }//rtc
+                            //if (intMetadataID == 4) {
+                            //    //this is the state - check for alarm
+                            //    SetControlText(this.meta_id_4,strValue);
+                                
+                            //}
+                            //if (intMetadataID == 1) { SetControlText(this.meta_id_1,strValue); }
                             
                             //leak sensors
                             if (intMetadataID == 2) { 
-                                SetControlText(this.meta_id_2, strValue);
+                                //SetControlText(this.meta_id_2, strValue);
                                 if (strValue == "1")
                                 {AlertUser(true);} 
                                 else { AlertUser(false); }
                                 
                             }
                             if (intMetadataID == 3) {
-                                SetControlText(this.meta_id_3, strValue);
+                                //SetControlText(this.meta_id_3, strValue);
                                 if (strValue == "1")
                                 {AlertUser(true);}
                                 else { AlertUser(false); }   
                             }
-                            if (intMetadataID == 7) { SetControlText(this.meta_id_7,strValue); }
-                            if (intMetadataID == 10) { SetControlText(this.meta_id_10,strValue); }
-                            if (intMetadataID == 11) { SetControlText(this.meta_id_11,strValue); }
-                            if (intMetadataID == 17) { SetControlText(this.meta_id_17, strValue); }
-                            if (intMetadataID == 18) { SetControlText(this.meta_id_18, strValue); }
-                            if (intMetadataID == 19) { SetControlText(this.meta_id_19, strValue); }
-                            if (intMetadataID == 21) { SetControlText(this.meta_id_21, strValue); }
-                            if (intMetadataID == 14) { SetControlText(this.meta_id_14, strValue); }
-                            if (intMetadataID == 15) { SetControlText(this.meta_id_15, strValue); }
-                            if (intMetadataID == 16) { SetControlText(this.meta_id_16, strValue); }
-                            if (intMetadataID == 22) { SetControlText(this.meta_id_22, strValue); }
-                            if (intMetadataID == 23) { SetControlText(this.meta_id_23, strValue); }
-                            if (intMetadataID == 24) { SetControlText(this.meta_id_24, strValue); }
-                            if (intMetadataID == 25) { SetControlText(this.meta_id_25, strValue); }
-                            if (intMetadataID == 26) { SetControlText(this.meta_id_26, strValue); }
+                            //if (intMetadataID == 7) { SetControlText(this.meta_id_7,strValue); }
+                            //if (intMetadataID == 10) { SetControlText(this.meta_id_10,strValue); }
+                            //if (intMetadataID == 11) { SetControlText(this.meta_id_11,strValue); }
+                            //if (intMetadataID == 17) { SetControlText(this.meta_id_17, strValue); }
+                            //if (intMetadataID == 18) { SetControlText(this.meta_id_18, strValue); }
+                            //if (intMetadataID == 19) { SetControlText(this.meta_id_19, strValue); }
+                            //if (intMetadataID == 21) { SetControlText(this.meta_id_21, strValue); }
+                            //if (intMetadataID == 14) { SetControlText(this.meta_id_14, strValue); }
+                            //if (intMetadataID == 15) { SetControlText(this.meta_id_15, strValue); }
+                            //if (intMetadataID == 16) { SetControlText(this.meta_id_16, strValue); }
+                            //if (intMetadataID == 22) { SetControlText(this.meta_id_22, strValue); }
+                            //if (intMetadataID == 23) { SetControlText(this.meta_id_23, strValue); }
+                            //if (intMetadataID == 24) { SetControlText(this.meta_id_24, strValue); }
+                            //if (intMetadataID == 25) { SetControlText(this.meta_id_25, strValue); }
+                            //if (intMetadataID == 26) { SetControlText(this.meta_id_26, strValue); }
+                            //if (intMetadataID == 32) { SetControlText(this.meta_id_32, strValue); }
+                            //if (intMetadataID == 33) { SetControlText(this.meta_id_33, strValue); }
+                            //if (intMetadataID == 34) { SetControlText(this.meta_id_34, strValue); }
 
                         }
 
@@ -218,8 +229,6 @@ namespace HarmoniaRemote
             sp.WriteLine("INFLATE,255");
         }
 
-        
-
         private void label1_Click(object sender, EventArgs e)
         {
 
@@ -241,8 +250,7 @@ namespace HarmoniaRemote
 
         private void tbFwdDive_Scroll(object sender, EventArgs e)
         {
-            sp.WriteLine("SERVOFWDDIVE," + this.tbFwdDive.Value.ToString());
-            
+            sp.WriteLine("SERVOFWDDIVE," + this.tbFwdDive.Value.ToString());        
         }
 
         private void btnIdle_Click(object sender, EventArgs e)
@@ -294,18 +302,29 @@ namespace HarmoniaRemote
 
         private void btnUpload_Click(object sender, EventArgs e)
         {
+            //do checks
+            if (txtLogPath.Text.Length == 0)
+            {
+                MessageBox.Show("Log path has not been set", "Settings Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
+            if (File.Exists(txtLogPath.Text))
+            {
+                MessageBox.Show("The log file already exists", "Upload Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                return;
+            }
+
             if (MessageBox.Show("Putting the system into Upload pause operation and upload SD Card data to a file - do you want to continue?","Continue?",MessageBoxButtons.OKCancel,MessageBoxIcon.Question) == DialogResult.OK)
             {
                 //have the file ready to go
-                m_strUploadFile = @"";
+                m_strUploadFile = txtLogPath.Text;
                 m_swLog = new System.IO.StreamWriter(m_strUploadFile, true);
 
                 m_blnUploading = true;
 
                 //change state to upload
                 sp.WriteLine("UPLOAD,0");
-                
-              
+                       
             }
          
         }
@@ -315,12 +334,36 @@ namespace HarmoniaRemote
             try
             {
 
-                NpgsqlConnection connPG = new NpgsqlConnection("");
+                //do checks
+                if (txtDBConn.Text.Length == 0)
+                {
+                    MessageBox.Show("DT database connection has not been set", "Settings Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
+                if (txtLogPath.Text.Length == 0)
+                {
+                    MessageBox.Show("Log path has not been set", "Settings Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
+                if (!File.Exists(txtLogPath.Text))
+                {
+                    MessageBox.Show("The log file path does not exist", "Settings Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
+                string strOutCSVFile = Path.GetFileNameWithoutExtension(txtLogPath.Text) + ".csv";
+                if (File.Exists(strOutCSVFile))
+                {
+                    MessageBox.Show("An output file already exists with the name: " + strOutCSVFile + " (this file may have already been converted)", "Conversion Error", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
+
+                //do the connect
+                NpgsqlConnection connPG = new NpgsqlConnection(txtDBConn.Text);
                 connPG.Open();
 
+                //get the metadata records
                 NpgsqlCommand commPG = new NpgsqlCommand("select metadata_id,data_label from dt_data_config order by metadata_id", connPG);
                 NpgsqlDataReader readerPG = commPG.ExecuteReader();
-
                 Hashtable hashDataLabels = new Hashtable();
                 while (readerPG.Read())
                 {
@@ -331,16 +374,15 @@ namespace HarmoniaRemote
                 readerPG.Close();
                 commPG.Dispose();
                 connPG.Close();
+  
+                //open the logfile
+                StreamReader reader = File.OpenText(txtLogPath.Text);
+                
+                //open the output csv file
+                StreamWriter swOut = new System.IO.StreamWriter(strOutCSVFile, false);
 
-
-                StreamWriter swOut = new System.IO.StreamWriter(@"", false);
-                StreamReader reader = File.OpenText(@"");
-                DateTime dteCurrent = DateTime.ParseExact("17:45:0 6/2/2023", "H:m:s d/M/yyyy", null);
-                DateTime dtePrevious = DateTime.Now;
                 int intRecord = 1;
-
                 String strFirstHeader = "";
-
                 while (reader.Peek() != -1)
                 {
 
@@ -350,9 +392,7 @@ namespace HarmoniaRemote
 
                     string strCSV = "";
                     string strComma = "";
-
                     string strHeader = "";
-
 
                     foreach (string strDataValue in arrayLine)
                     {
@@ -375,30 +415,6 @@ namespace HarmoniaRemote
                                                                                                           //DateTime dteUTM = dteLoc.ToUniversalTime();
                                 strValue = dteThis.ToString();
 
-                                //if (intRecord > 1)
-                                //{
-                                //    TimeSpan ts = dteThis - dtePrevious;
-                                //    if (ts.TotalSeconds >= 0)
-                                //    {
-                                //        //Console.WriteLine(ts.TotalSeconds.ToString() + " gap at " + dteThis.ToString());
-                                //        //MessageBox.Show(dteThis.ToString());
-
-                                //        dteCurrent = dteCurrent.AddSeconds(ts.TotalSeconds);
-
-                                //    }
-                                //    else
-                                //    {
-                                //        //time has been reset - just add a 5 minute can because sub was probably turned off
-                                //        dteCurrent = dteCurrent.AddMinutes(5);
-                                //    }
-                                //}
-
-                                ////Console.WriteLine(dteThis.ToString() + ">>" + dteCurrent.ToString());
-
-                                //strValue = dteCurrent.ToString();
-
-
-                                //dtePrevious = dteThis;
                             }
 
                             strHeader = strHeader + strComma + strLabel;
